+++ packages/multichain-account-service/src/MultichainAccountService-method-action-types.ts
@@ -6,6 +6,15 @@
 import type { MultichainAccountService } from './MultichainAccountService';
 
 /**
+ * Initialize the service and constructs the internal reprensentation of
+ * multichain accounts and wallets.
+ */
+export type MultichainAccountServiceInitAction = {
+  type: `MultichainAccountService:init`;
+  handler: MultichainAccountService['init'];
+};
+
+/**
  * Re-synchronize MetaMask accounts and the providers accounts if needed.
  *
  * NOTE: This is mostly required if one of the providers (keyrings or Snaps)
@@ -23,11 +32,6 @@ import type { MultichainAccountService } from './MultichainAccountService';
 export type MultichainAccountServiceResyncAccountsAction = {
   type: `MultichainAccountService:resyncAccounts`;
   handler: MultichainAccountService['resyncAccounts'];
-};
-
-export type MultichainAccountServiceEnsureCanUseSnapPlatformAction = {
-  type: `MultichainAccountService:ensureCanUseSnapPlatform`;
-  handler: MultichainAccountService['ensureCanUseSnapPlatform'];
 };
 
 /**
@@ -190,8 +194,8 @@ export type MultichainAccountServiceAlignWalletAction = {
  * Union of all MultichainAccountService action types.
  */
 export type MultichainAccountServiceMethodActions =
+  | MultichainAccountServiceInitAction
   | MultichainAccountServiceResyncAccountsAction
-  | MultichainAccountServiceEnsureCanUseSnapPlatformAction
   | MultichainAccountServiceGetMultichainAccountWalletAction
   | MultichainAccountServiceGetMultichainAccountWalletsAction
   | MultichainAccountServiceCreateMultichainAccountWalletAction